Refactors header component in forms #1760
