+++ src/components/HeaderWithLanguage/index.js
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2019-present, NDLA.
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import HeaderWithLanguage from './HeaderWithLanguage';
+
+export default HeaderWithLanguage;
